server/requirements.txt
Django>=4.0
psycopg2>=2.8
djoser
django-filter
djangorestframework
django-cors-headers
django-extensions
scrapy
scrapy_djangoitem
scrapy-inline-requests
djangorestframework-simplejwt
Pillow
django-environ
sendgrid-django
django-crontab